// src/interfaces/Venta.interfaces.tsx
export interface Venta {
    codVenta:            number;
    codEmpresa:          number;
    codSucursal:         number;
    codEmpresaErp:       string;
    codSucursalErp:      string;
    formaVenta:          FormaVenta;
    cliente:             Cliente;
    listaPrecio:         ListaPrecio;
    codVendedorErp:      string;
    fechaVenta:          string;
    fechaCreacion:       string;
    fechaModificacion:   string;
    codUsuarioCreacion:  number;
    terminalVenta:       TerminalVenta;
    tipoComprobante:     string;
    nroComprobante:      string;
    timbrado:            string;
    inicioTimbrado:      string;
    finTimbrado:         string;
    fechaVencimiento:    string;
    importeTotal:        number;
    subTotal:            number;
    importeNeto:         number;
    porcDescuento:       number;
    importeDescuento:    number;
    importeIva5:         number;
    importeIva10:        number;
    importeIvaExenta:    number;
    totalKg:             number;
    anulado:             boolean;
    esObsequio:          boolean;
    editable:            boolean;
    motivoAnulacion:     null;
    fechaAnulacion:      null;
    codUsuarioAnulacion: null;
    detalle:             VentaDetalle[];
    cobranza:            Cobranza;
    tipoVenta:           string;
    modoEntrega:         string;
    estado:              string;
    pedido:              null;
    cupon:               null;
    vendedor:            Vendedor;
    codReparto:          number;
    deposito:            Deposito;
    descuentoProducto:   number;
    puntosObtenidos:     number;
    totalPuntos:         number;
}

export interface Cliente {
    codCliente:             number;
    codClienteErp:          string;
    razonSocial:            string;
    tipoDoc:                string;
    docNro:                 string;
    codEmpresa:             number;
    sucursalPref:           Sucursal;
    listaPrecio:            ListaPrecio;
    medioPagoPref:          MedioPago;
    formaVentaPref:         FormaVenta;
    grupoDescuento:         GrupoDescuento;
    carnetGrupo:            string;
    carnetVencimiento:      string;
    categoria:              string;
    grupo:                  string;
    zona:                   string;
    direccion:              string;
    telefono:               string;
    obssap:                 null;
    email:                  string;
    web:                    string;
    catABC:                 string;
    obs:                    string;
    activo:                 boolean;
    codeBarra:              boolean;
    diasCredito:            number;
    fechaCreacion:          string;
    fechaModificacion:      string;
    codUsuarioCreacion:     null;
    codUsuarioModificacion: number;
    codUltimaVenta:         number;
    excentoIva:             boolean;
    predeterminado:         boolean;
    esPropietario:          boolean;
    empleado:               boolean;
    latitud:                number;
    longitud:               number;
    concatDocNombre:        string;
    concatCodNombre:        string;
    concatCodErpNombre:     string;
    puntos:                 number;
}

export interface FormaVenta {
    codFormaVenta:    number;
    descripcion:      string;
    cantDias:         number;
    codFormaVentaErp: string;
    codEmpresa:       number;
    esContado:        boolean;
}

export interface GrupoDescuento {
    codGrupo:    number;
    codEmpresa:  number;
    descripcion: string;
    descuento:   number;
}

export interface ListaPrecio {
    codListaPrecio:         number;
    codListaPrecioErp:      string;
    descripcion:            string;
    codEmpresa:             number;
    ecommerce:              boolean;
    sincronizable:          boolean;
    fechaCreacion:          null;
    fechaModificacion:      null;
    codUsuarioCreacion:     null;
    codUsuarioModificacion: null;
}

export interface MedioPago {
    codMedioPago:           number;
    codMedioPagoErp:        string;
    descripcion:            string;
    codEmpresa:             number;
    tieneTipo:              boolean;
    tieneRef:               boolean;
    tieneBanco:             boolean;
    esCheque:               boolean;
    esObsequio:             boolean;
    fechaCreacion:          null;
    fechaModificacion:      null;
    codUsuarioCreacion:     null;
    codUsuarioModificacion: null;
}

export interface Sucursal {
    codSucursal:            number;
    codSucursalErp:         string;
    centro:                 string;
    nombreSucursal:         string;
    direccion:              string;
    telefono:               string;
    email:                  string;
    principal:              boolean;
    codEmpresa:             number;
    latitud:                number;
    longitud:               number;
    modoVendedor:           string;
    fechaCreacion:          null;
    fechaModificacion:      null;
    codUsuarioCreacion:     null;
    codUsuarioModificacion: null;
    envioposventa:          boolean;
    mensaje:                string;
}

export interface Cobranza {
    codCobranza:         number;
    codEmpresa:          number;
    codSucursal:         number;
    codUsuarioCreacion:  number;
    fechaCobranza:       string;
    importeCobrado:      number;
    importeAbonado:      number;
    saldo:               number;
    anulado:             boolean;
    codDoc:              null;
    codUsuarioAnulacion: null;
    fechaAnulacion:      null;
    tipo:                string;
    detalle:             CobranzaDetalle[];
}

export interface CobranzaDetalle {
    codCobranzaDetalle: number;
    medioPago:          MedioPago;
    tipoMediopago:      null;
    importeCobrado:     number;
    importeAbonado:     number;
    saldo:              number;
    nroRef:             null;
    fechaEmision:       null;
    fechaVencimiento:   null;
    banco:              null;
    nroCuenta:          null;
}

export interface Deposito {
    codDeposito:    number;
    codDepositoErp: string;
    nombreDeposito: string;
    codEmpresa:     number;
    sucursal:       Sucursal;
    tipoDeposito:   TipoDeposito;
    tipoVenta:      boolean;
}

export interface TipoDeposito {
    codTipoDeposito: number;
    descripcion:     string;
    codEmpresa:      number;
}

export interface VentaDetalle {
    codVentaDetalle:  number;
    nroItem:          number;
    producto:         Producto;
    codVendedorErp:   string;
    unidadMedida:     Unidad;
    cantidad:         number;
    importePrecio:    number;
    importeTotal:     number;
    subTotal:         number;
    importeNeto:      number;
    porcDescuento:    number;
    importeDescuento: number;
    porcIva:          number;
    importeIva5:      number;
    importeIva10:     number;
    importeIvaExenta: number;
    totalKg:          number;
    vendedor:         Vendedor;
    tipoDescuento:    string;
}

export interface Producto {
    codProducto:            number;
    codProductoErp:         string;
    nombreProducto:         string;
    empresa:                Empresa;
    unidad:                 Unidad;
    descripcion:            string;
    codBarra:               string;
    categoriaProducto:      CategoriaProducto;
    marca:                  string;
    presentacion:           string;
    color:                  string;
    peso:                   number;
    catABC:                 string;
    img:                    string;
    obs:                    string;
    destacado:              boolean;
    inventariable:          boolean;
    activo:                 boolean;
    sinDescuento:           boolean;
    iva:                    number;
    ivaEspecial:            boolean;
    cantidadMin:            number;
    cantidadMax:            number;
    fechaCreacion:          null;
    fechaModificacion:      null | string;
    codUsuarioCreacion:     null;
    codUsuarioModificacion: number | null;
    grpMaterial:            null;
    concatCodNombre:        string;
    concatCodErpNombre:     string;
    codMaterial:            string;
}

export interface CategoriaProducto {
    codCategoriaProducto:    number;
    codCategoriaProductoErp: string;
    descripcion:             string;
    codEmpresa:              number;
}

export interface Empresa {
    codEmpresa:       number;
    codEmpresaErp:    string;
    razonSocial:      string;
    nombre:           string;
    actividad1:       string;
    actividad2:       string;
    direccionGen:     string;
    telefonoGen:      string;
    ruc:              string;
    tipoNegocio:      TipoNegocio;
    rubro:            null;
    img:              string;
    logoReporte:      null;
    logoHeaderDark:   null;
    logoHeaderLight:  null;
    logoTextDark:     null;
    logoTextLight:    null;
    cantItem:         number;
    maxDescuentoPorc: number;
    maxDescuentoImp:  number;
    tipoLicencia:     TipoLicencia;
}

export interface TipoLicencia {
    codTipoLicencia: number;
    descripcion:     string;
    cantTerminales:  number;
    cantSucursales:  number;
    cantUsuarios:    number;
    precio:          number;
}

export interface TipoNegocio {
    codTipoNegocio: number;
    descripcion:    string;
    valor:          string;
}

export interface Unidad {
    codUnidad:    number;
    codUnidadErp: string;
    nombreUnidad: string;
    codEmpresa:   number;
    cantidad:     number;
}

export interface Vendedor {
    codVendedor:        number;
    codVendedorErp:     string;
    vendedor:           string;
    docNro:             string;
    codEmpresa:         number;
    usuario:            Usuario;
    direccion:          string;
    telefono:           string;
    email:              string;
    tarjeta:            null;
    codSupervisorErp:   string;
    tipo:               string;
    porcentajeComision: number;
    obs:                string;
    activo:             boolean;
    fechaCreacion:      string;
    fechaModificacion:  string;
    concatDocNombre:    string;
    concatCodErpNombre: null;
}

export interface Usuario {
    codUsuario:     number;
    nombrePersona:  string;
    codPersonaErp:  string;
    username:       string;
    rol:            Rol;
    codEmpresa:     number;
    sucursal:       Sucursal;
    createdAt:      string;
    modifiedAt:     string;
    lastLoginAt:    string;
    createdBy:      string;
    modifiedBy:     string;
    enabled:        boolean;
    img:            string;
    bloqueado:      boolean;
    intentoFallido: number;
}

export interface Rol {
    codRol:    number;
    nombreRol: string;
}

export interface TerminalVenta {
    codTerminalVenta: number;
    descripcion:      string;
    id:               string;
    codEmpresa:       number;
    codSucursal:      number;
    disponible:       boolean;
}